Implemented scalar copying for JikesRVM object model

+++ src/vm/jikesrvm/java_header.rs
@@ -7,7 +7,7 @@ pub const SCALAR_HEADER_SIZE: usize = JAVA_HEADER_BYTES + OTHER_HEADER_BYTES;
 pub const ARRAY_HEADER_SIZE: usize = SCALAR_HEADER_SIZE + ARRAY_LENGTH_BYTES;
 
 /** offset of object reference from the lowest memory word */
-pub const OBJECT_REF_OFFSET: usize = ARRAY_HEADER_SIZE;  // from start to ref
+pub const OBJECT_REF_OFFSET: isize = ARRAY_HEADER_SIZE as isize;  // from start to ref
 pub const TIB_OFFSET: isize = JAVA_HEADER_OFFSET;
 pub const STATUS_OFFSET: isize = TIB_OFFSET + STATUS_BYTES as isize;
 #[cfg(target_endian = "little")]
